move authentication logic to helpers file
