+++ app/controllers/api/v1/base.rb
@@ -1,6 +1,7 @@
 module API
   module V1
     class Base < Grape::API
+      helpers API::AuthHelpers
       mount API::V1::Users
       mount API::V1::Sessions
       mount API::V1::Vaults
